requirements.txt
requests==2.32.3
tqdm==4.64.1
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
